requirements.txt
accelerate==0.32.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
boilerpy3==1.0.7
boto3==1.34.143
botocore==1.34.143
build==1.2.1
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
databricks-api==0.9.0
databricks-cli==0.18.0
dataclasses==0.6
dataclasses-json==0.6.7
decorator==5.1.1
deepdiff==7.0.1
deepmultilingualpunctuation==1.0.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
effdet==0.4.1
emoji==2.12.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.3
filelock==3.15.4
filetype==1.2.0
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.30.0
google-cloud-vision==3.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
InstructorEmbedding==1.0.1
intel-openmp==2021.4.0
iopath==0.1.10
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
johnsnowlabs==5.3.6
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.2.6
langchain-cli==0.0.25
langchain-community==0.2.6
langchain-core==0.2.10
langchain-text-splitters==0.2.2
langcodes==3.4.0
langdetect==1.0.9
langserve==0.2.2
langsmith==0.1.82
language_data==1.2.0
layoutparser==0.3.4
libcst==1.4.0
llama-cloud==0.0.6
llama-index-core==0.10.51
llama-index-embeddings-openai==0.1.10
llama-index-readers-file==0.1.26
lxml==5.2.2
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
nlu==5.3.2
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.18.1
openai==1.35.9
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
ordered-set==4.1.0
orjson==3.10.5
overrides==7.7.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.1
pikepdf==9.0.0
pillow==10.3.0
pillow_heif==0.16.0
playwright==1.45.0
plum-dispatch==1.7.4
poppler-utils==0.1.0
portalocker==2.10.0
posthog==3.5.0
preshed==3.0.9
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
psycopg2==2.9.9
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyee==11.1.0
Pygments==2.18.0
PyJWT==2.8.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.2.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject-toml==0.0.10
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pyspark==3.4.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.9
python-oxmsg==0.0.1
python-pptx==0.6.23
pytz==2024.1
pywin32==306
PyYAML==6.0.1
rapidfuzz==3.9.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-splitter==1.4
sentence-transformers==2.2.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spark-nlp==5.3.2
spark-nlp-display==5.0
Spire.Pdf==10.6.1
SQLAlchemy==2.0.31
srsly==2.4.8
sse-starlette==1.8.2
stack-data==0.6.3
stanza==1.8.2
starlette==0.37.2
striprtf==0.0.26
svgwrite==1.4
sympy==1.12.1
tabulate==0.9.0
tbb==2021.13.0
tenacity==8.4.2
textblob==0.18.0.post0
thefuzz==0.22.1
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.7
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.5
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers.git@d625294d79341662784495551abdf45e6cb9372f
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.14.9
unstructured-client==0.23.8
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.12
urllib3==2.2.2
uvicorn==0.23.2
wasabi==1.1.3
watchfiles==0.22.0
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.19.2
